% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_family_data.R
\name{FT_SEM_process_data function}
\alias{FT_SEM_process_data function}
\title{The main function for family trios structural equation model (FT-SEM)}
\arguments{
\item{data}{A data.frame about genotype from plink1.9, containing FID,IID,PAT,MAT,SEX, phnotype and genotype (only 1 coloum)}

\item{seed}{random seed for select offspring from a family with multiple offsprings}
}
\description{
A robust and powerful GWAS method for family trios
}
