% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_family_data.R
\name{process_family_data}
\alias{process_family_data}
\title{The main function for family trios structural equation model (FT-SEM)}
\usage{
process_family_data(data, seed = 123)
}
\arguments{
\item{data}{A data.frame about genotype from plink1.9, containing FID,IID,PAT,MAT,SEX, phnotype and genotype (only 1 coloum)}

\item{seed}{random seed for select offspring from a family with multiple offsprings}
}
\description{
A robust and powerful GWAS method for family trios
}
